+++ src/ui/alert.rs
@@ -1,7 +1,8 @@
 use crate::ui::alert::AlertResult::{Negative, Positive};
-use crate::ui::prelude::Positioning::Center;
 use crate::ui::prelude::*;
 use crate::ui::styles::AlertStyle;
+use buffer_graphics_lib::prelude::Positioning::Center;
+use buffer_graphics_lib::prelude::*;
 use graphics_shapes::coord;
 
 const BUTTON_Y: isize = 28;