+++ src/ui/styles/defaults.rs
@@ -1,5 +1,6 @@
-use crate::prelude::styles::*;
 use crate::prelude::MIN_FILE_DIALOG_SIZE;
+use crate::prelude::*;
+use crate::ui::prelude::*;
 use buffer_graphics_lib::prelude::Positioning::LeftTop;
 use buffer_graphics_lib::prelude::*;
 
@@ -172,7 +173,7 @@ impl Default for ToggleIconButtonStyle {
             shadow: ToggleColorSet::new_values(
                 DARK_GRAY, DARK_GRAY, WHITE, WHITE, DARK_GRAY, DARK_GRAY,
             ),
-            rounding: 4,
+            rounding: 6,
             padding: 4,
         }
     }