+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### Version 0.11.4
+- Add `set_position` to `UiElement` and all views
+- Add `ColumnLayout` and `RowLayout` which can position a collection of views into columns or rows
+
 ### Version 0.11.3
 - Update buffer/ici-files